feat(queue_manager): Include queue type in queue name
When changing the queue type of a client queue, its next declaration is
going to fail with error code 406 (Precondition failed), since it does
not match the arguments of the existing queue.

We include the queue type in the queue name to sidestep the problem.
Changing the queue type then results in a new queue (with a predictable
name) being declared and no conflicts arise.  A client re-registering
itself with a changed queue type then gets assigned the newly declared
queue.  Of course, the old queue still exists and needs to be deleted
manually.

For compatibility with the old code, wee keep queue old names for
default queue type.